requirements.txt
SQLAlchemy==1.3.12
cryptography==2.8
PyQRCode==1.2.1
Pillow==6.1.0
pypng==0.0.20
pyperclip==1.7.0
termcolor==1.1.0